requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==5.1.4
django-admin-inline-paginator==0.4.0
django-admin-rangefilter==0.12.0
django-ckeditor==6.6.1
django-cors-headers==4.3.0
django-js-asset==2.1.0
django-rest-swagger==2.2.0
django-storages==1.14.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
openapi-codec==1.3.2
psycopg2==2.9.9
pycosat==0.6.6
PyJWT==2.10.1
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
simplejson==3.19.3
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
